refactor: reorganize dependency injection and application services
