+++ src/backend/noKeyCloud.Application/DependencyInjection.cs
@@ -1,0 +1,17 @@
+﻿using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace noKeyCloud.Application;
+
+public static class DependencyInjection
+{
+    /// <summary>
+    /// Configures application layer services, including MediatR for CQRS and other application-specific services(Validators, AutoMapper, etc.).
+    /// </summary>
+    public static IServiceCollection AddApplication(this IServiceCollection services)
+    {
+        services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly()));
+
+        return services;
+    }
+}
